+++ PHP/CodeCoverage/Autoload.php
@@ -59,6 +59,7 @@ spl_autoload_register(
             'php_codecoverage_driver' => '/CodeCoverage/Driver.php',
             'php_codecoverage_driver_xdebug' => '/CodeCoverage/Driver/Xdebug.php',
             'php_codecoverage_exception' => '/CodeCoverage/Exception.php',
+            'php_codecoverage_exception_unintentionallycoveredcode' => '/CodeCoverage/Exception/UnintentionallyCoveredCode.php',
             'php_codecoverage_filter' => '/CodeCoverage/Filter.php',
             'php_codecoverage_report_clover' => '/CodeCoverage/Report/Clover.php',
             'php_codecoverage_report_factory' => '/CodeCoverage/Report/Factory.php',