+++ ChangeLog.markdown
@@ -6,5 +6,6 @@ This is the list of changes for the PHP_CodeCoverage 1.3 release series.
 PHP_CodeCoverage 1.3.0
 ----------------------
 
+* Added an option to raise an exception when code is unintentionally covered.
 * PHP_CodeCoverage 1.3 is only supported on PHP 5.4.7 (or later).